Extract enterprise request context module
